de-fatal state serialisation and remove obsolete state-functions

+++ ssh/kex.h
@@ -106,7 +106,7 @@ struct ssh;
 
 struct Kex {
 	u_char	*session_id;
-	u_int	session_id_len;
+	size_t	session_id_len;
 	Newkeys	*newkeys[MODE_MAX];
 	u_int	we_need;
 	int	server;
